requirements.txt
Flask==3.1.0
flask-cors==5.0.1
openai
python-dotenv==1.1.0
gunicorn==20.1.0
requests==2.32.3